+++ wled00/wled.cpp
@@ -264,15 +264,15 @@ void WLED::initAP(bool resetAP)
     return;
 
   if (!apSSID[0] || resetAP)
-    strcpy(apSSID, (const char*)F("WLED-AP"));
+    strcpy_P(apSSID, PSTR("WLED-AP"));
   if (resetAP)
-    strcpy(apPass, DEFAULT_AP_PASS);
+    strcpy_P(apPass, PSTR(DEFAULT_AP_PASS));
   DEBUG_PRINT(F("Opening access point "));
   DEBUG_PRINTLN(apSSID);
   WiFi.softAPConfig(IPAddress(4, 3, 2, 1), IPAddress(4, 3, 2, 1), IPAddress(255, 255, 255, 0));
   WiFi.softAP(apSSID, apPass, apChannel, apHide);
 
-  if (!apActive)        // start captive portal if AP active
+  if (!apActive) // start captive portal if AP active
   {
     DEBUG_PRINTLN(F("Init AP interfaces"));
     server.begin();
@@ -283,6 +283,17 @@ void WLED::initAP(bool resetAP)
       udpRgbConnected = rgbUdp.begin(udpRgbPort);
     }
 
+    if (udpPort2 > 0 && udpPort2 != ntpLocalPort && udpPort2 != udpPort && udpPort2 != udpRgbPort) {
+      udp2Connected = notifier2Udp.begin(udpPort2);
+    }
+    
+    if (audioSyncPort > 0 || (((audioSyncEnabled)>>(0)) & 1) || (((audioSyncEnabled)>>(1)) & 1)) {
+    #ifndef ESP8266
+      udpSyncConnected = fftUdp.beginMulticast(IPAddress(239,0,0,1), audioSyncPort);
+    #else
+      udpSyncConnected = fftUdp.beginMulticast(WiFi.localIP(), IPAddress(239, 0, 0, 1), audioSyncPort);
+    #endif
+    }
     dnsServer.setErrorReplyCode(DNSReplyCode::NoError);
     dnsServer.start(53, "*", WiFi.softAPIP());
   }
@@ -353,7 +364,7 @@ void WLED::initConnection()
       pos--;
     }
   }
-  
+
 #ifdef ESP8266
   WiFi.hostname(hostname);
 #endif
@@ -408,6 +419,15 @@ void WLED::initInterfaces()
     udpConnected = notifierUdp.begin(udpPort);
     if (udpConnected && udpRgbPort != udpPort)
       udpRgbConnected = rgbUdp.begin(udpRgbPort);
+    if (udpConnected && udpPort2 != udpPort && udpPort2 != udpRgbPort)
+      udp2Connected = notifier2Udp.begin(udpPort2);
+  }
+  if (audioSyncPort > 0 || (((audioSyncEnabled)>>(0)) & 1) || (((audioSyncEnabled)>>(1)) & 1)) {
+    #ifndef ESP8266
+      udpSyncConnected = fftUdp.beginMulticast(IPAddress(239,0,0,1), audioSyncPort);
+    #else
+      udpSyncConnected = fftUdp.beginMulticast(WiFi.localIP(), IPAddress(239, 0, 0, 1), audioSyncPort);
+    #endif
   }
   if (ntpEnabled)
     ntpConnected = ntpUdp.begin(ntpLocalPort);